Fixed 'jsx' in '.js' file
