+++ app/javascript/src/components/Dashboard/Contacts/constants.js
@@ -1,16 +1,4 @@
-import React from "react";
-
 import * as yup from "yup";
-
-import Avatar from "./Table/Avatar";
-import RowActions from "./Table/RowActions";
-import { convertDateToStdFormat } from "./utils";
-
-const DUMMY_CONTACT_DATA = {
-  name: "Ronald Reagan",
-  email: "ronald@gmail.com",
-  createdAt: convertDateToStdFormat(new Date()),
-};
 
 export const ROLES = [
   "UX designer",
@@ -43,35 +31,3 @@ export const CONTACT_FORM_VALIDATION_SCHEMA = yup.object().shape({
     })
     .required("Role is required"),
 });
-
-export const ROW_DATA = Array(100).fill(DUMMY_CONTACT_DATA);
-
-export const COLUMN_DATA = [
-  {
-    title: "NAME & ROLE",
-    dataIndex: "name",
-    key: "name",
-    width: "33%",
-    // eslint-disable-next-line react/jsx-filename-extension
-    render: name => <Avatar name={name} />,
-  },
-  {
-    title: "EMAIL",
-    dataIndex: "email",
-    key: "email",
-    width: "28%",
-  },
-  {
-    title: "CREATED AT",
-    dataIndex: "createdAt",
-    key: "createdAt",
-    width: "25%",
-  },
-  {
-    title: "",
-    dataIndex: "icon_button",
-    key: "icon_button",
-    width: "20%",
-    render: () => <RowActions />,
-  },
-];
